+++ backend/models/sale_item.py
@@ -1,5 +1,6 @@
 import uuid
-from sqlalchemy import Float, ForeignKey
+from decimal import Decimal
+from sqlalchemy import Numeric, ForeignKey
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -15,8 +16,9 @@ class SaleItem(Base):
         UUID(as_uuid=True), ForeignKey("sales.id"))
     product_id: Mapped[uuid.UUID] = mapped_column(
         UUID(as_uuid=True), ForeignKey("products.id"))
-    quantity: Mapped[float] = mapped_column(Float, nullable=False)
-    price_at_sale: Mapped[float] = mapped_column(Float, nullable=False)
+    quantity: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
+    price_at_sale: Mapped[Decimal] = mapped_column(
+        Numeric(10, 2), nullable=False)
 
     sale = relationship("Sale", back_populates="items")
     product = relationship("Product", back_populates="sale_items")
